refactor(*): update references to dependency type
