+++ src/decorators/js-service.decorator.ts
@@ -1,4 +1,4 @@
-import { AnyServiceDependency } from '../interfaces/service-options-dependency.interface';
+import { AnyServiceDependency } from '../interfaces/service-dependency.interface';
 import { ServiceOptions } from '../interfaces/service-options.interface';
 import { AnyConstructable } from '../types/any-constructable.type';
 import { AnyInjectIdentifier } from '../types/inject-identifier.type';
