+++ src/decorators/service.decorator.ts
@@ -14,7 +14,7 @@ import { ServiceMetadata } from '../interfaces/service-metadata.interface';
 import { SERVICE_METADATA_DEFAULTS } from '../constants/service-defaults.const';
 import { BUILT_INS } from '../constants/builtins.const';
 import { CannotInstantiateBuiltInError } from '../error/cannot-instantiate-builtin-error';
-import { AnyServiceDependency } from '../interfaces/service-options-dependency.interface';
+import { AnyServiceDependency } from '../interfaces/service-dependency.interface';
 import { wrapDependencyAsResolvable } from '../utils/wrap-resolvable-dependency';
 import { throwError } from '../utils/throw-error.util';
